(fix) Restructure to do processing in jobs rather than inline
