+++ imports/plugins/included/connectors-shopify/server/index.js
@@ -3,3 +3,4 @@ import "./i18n";
 import "./methods";
 import "./endpoints/webhooks";
 import "./hooks";
+import "./jobs";
