+++ imports/plugins/included/connectors-shopify/server/jobs/index.js
@@ -1,0 +1,3 @@
+import exportOrders from "./order-export";
+
+exportOrders();
