+++ imports/plugins/included/connectors-shopify/server/methods/index.js
@@ -1,4 +1,5 @@
 import "./api";
 import "./import";
+import "./export";
 import "./sync";
 import "./webhooks";
